requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==3.2.20
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.4
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==1.26.7
